Add compiled versions of cilantro and harvest js/css
Our previous strategy was to pull the latest version of Cilantro from git and compile it. In favor of dropping as many dependencies as possible the harvest cli utility now includes precompiled versions of the Cilantro thin-client as well as compiled versions of the Harvest app's coffescript and sass files.